Add typing hints
This is an attempt at add type hints to the codebase.

I'm mostly doing this since I want to better understand some of the
helper functions so as to expose a few and make them reusable.

There's still a issues in `_format_envvars` and `_generate_nodes`. Both
of them take an argument and pass it to a conflicting function.