Add ":hide-header:", ":nested: complete"
Adding :hide-header: to the directive will hide the title, description
and usage and only show the commands. This also means that the commands
will not be under a single section, but rather add a section each to
where the directive is located.

The ":nested: complete" option shows the name, description, usage, short
command overview (basically the entire ":nested: short" option) followed
by the details of the commands from the ":nested: full" option.