Add anchors to options, arguments and envvars
Add anchors to all the Click components (i.e. groups and commands),
along with another general one for options so that now using the full
path of the component with separating dashes is always a valid
reference. E.g. for a group "mygroup", command "mycommand" and option
"-myoption", all of the following work:

  :ref:`My Group <mygroup>`
  :ref:`My Command <mygroup-mycommand>`
  :ref:`My Option <mygroup-mycommand-myoption>`
  :option:`My Option <mygroup-mycommand -myoption>`