FIX Format string defaults correctly
Before this patch, string defaults are formatted without quotes around
them. However, if the default is a tuple or list of strings, quotes will
be included:

  default="a" ==> :default: ``a``
  default=("a",) ==> :default: ``('a')``

This is a slightly annoying inconsistency, and leads to potential for
confusion between `default="0"` and `default=0`. Most bothersome is that
in the case of an empty string we get:

  :default: ````

Which makes docutils angry:

  CRITICAL: Unexpected section title or transition.

This fixes the trouble by formatting the repr of the default.